refactor benchmark module to allow extension with performance benchmarks
